Add a word of caution to spread operator use
We thought the spread operator was very nifty and cool in our UI library until we discovered we had invalid HTML attributes all over the place. Now it's pretty much banned from the library, except for instances where the props are exactly the same (such as Icon props get passed into an iconButton which is an icon with an `onClick` callback).